Archi and Comparsion
Danya gave integers a, b and n to Archi. Archi wants to compare an and bn

. Help Archi with this task.
Input

    The first line of the input contains a single integer T

denoting the number of test cases. The description of T
test cases follows.
The first and only line of each test case contains three space-separated integers a
, b and n

    .

Output

For each test case, print a single line containing one integer: 1
if an>bn, 2 if an<bn or 0 if an=bn

.
Constraints

    1≤T≤1,000

|a|,|b|≤109
1≤n≤109

Example Input

2
3 4 5
-3 2 4

Example Output

2
1

Explanation

Example case 1: an=243
and bn=1024

.

Example case 2: an=81
and bn=16.